Big update with most of the code.
This is the first real update with the NextJS project and many
components. In particular we have Admin component for the page layout,
FileList component for visualizing the list of files, and SwitchWrapper
for the dark mode toggler. This is still a work-in-progress and should
be considered experimental, thus not yet usable.